requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas>=2.0.0
openpyxl>=3.1.2
SQLAlchemy>=2.0.23
numpy>=1.26.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
